% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquaire.R
\name{rsquaire}
\alias{rsquaire}
\title{rsquaire}
\usage{
rsquaire(data, index = NULL, indexType = "numeric", categories = NULL,
  layout = NULL, labels = NULL, labelStyle = "short",
  colors = c("#CFF09E", "#A8DBA8", "#79BD9A", "#3B8686", "#0B486B"),
  tooltip = FALSE, column1 = "", column2 = "", whitelist = NULL,
  mode = "dynamic", noteIndex = NULL, width = NULL, height = NULL,
  elementId = "map-container")
}
\arguments{
\item{data}{A dataframe in "wide format." One variable of the dataframe will need to be named
'state' and contain state abbrevations. Each other variable is a statistic about that state.}

\item{index}{A string which tells rsquaire which variable to use for the map coloring}

\item{labelStyle}{A string indicating which type of label to use on the map. Default options are
"short", "full," and "ap."}

\item{colors}{A character vector of color hex codes you want to use for map coloring}

\item{tooltip}{A boolean indicating if you want responsive tooltips enabled}

\item{column1}{A string to label the first column of your tooltip}

\item{column2}{A string to label the second column of your tooltip}

\item{whitelist}{A character vector of column names from your dataframe to include in your tooltip}

\item{mode}{A string indicating which type of tooltip you want}
}
\description{
Produces an htmlwidget of a equal-area square cartogram
}
